Move store to another folder
