+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.scss';
 import { Provider } from 'react-redux';
-import { store } from './app/store';
+import { store } from './store';
 import Routes from './Routes';
 
 ReactDOM.render(
